+++ tests/findex/findex.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jsdoc/require-jsdoc */
 import {
   FetchChains,
   FetchEntries,
@@ -19,27 +18,101 @@ import {
 import { USERS } from "../data/users"
 import { expect, test } from "@jest/globals"
 import { createClient } from "redis"
+import { hexEncode } from "../../src/utils/utils"
+import { randomBytes } from "crypto"
+import sqlite3 = require("sqlite3")
 
-test("upsert in memory", async () => {
+
+test("upsert and search memory", async () => {
   const entryLocation: IndexedEntry = {
-    indexedValue: IndexedValue.fromLocation(
-      new Location(new TextEncoder().encode("ROBERT file"))
-    ),
-    keywords: new Set([new Keyword(new TextEncoder().encode("ROBERT"))]),
+    indexedValue: IndexedValue.fromLocation(new Location(new TextEncoder().encode("ROBERT file"))),
+    keywords: new Set([new Keyword(new TextEncoder().encode("ROBERT"))])
   }
   const entryLocation_ = new LocationIndexEntry("ROBERT file", ["ROBERT"])
   expect(entryLocation_).toEqual(entryLocation)
 
   const entryKeyword: IndexedEntry = {
-    indexedValue: IndexedValue.fromNextWord(
-      new Keyword(new TextEncoder().encode("ROBERT"))
-    ),
-    keywords: new Set([new Keyword(new TextEncoder().encode("BOB"))]),
+    indexedValue: IndexedValue.fromNextWord(new Keyword(new TextEncoder().encode("ROBERT"))),
+    keywords: new Set([new Keyword(new TextEncoder().encode("BOB"))])
   }
   const entryKeyword_ = new KeywordIndexEntry("BOB", "ROBERT")
   expect(entryKeyword_).toEqual(entryKeyword)
 
-  // const label = new Label(new TextEncoder().encode("Q1 2022"))
+  const searchKey = new FindexKey(randomBytes(32))
+  const updateKey = new FindexKey(randomBytes(32))
+
+  const label = new Label("test")
+
+  const entryTable: { [uid: string]: Uint8Array } = {}
+  const chainTable: { [uid: string]: Uint8Array } = {}
+
+  const fetchEntries: FetchEntries = async (uids: Uint8Array[]): Promise<UidsAndValues> => {
+    const results: UidsAndValues = []
+    for (const uid of uids) {
+      const value = entryTable[hexEncode(uid)]
+      if (typeof value !== "undefined") {
+        results.push({ uid, value })
+      }
+    }
+    return await Promise.resolve(results)
+  }
+
+  const fetchChains: FetchChains = async (uids: Uint8Array[]): Promise<UidsAndValues> => {
+    const results: UidsAndValues = []
+    for (const uid of uids) {
+      const value = chainTable[hexEncode(uid)]
+      if (typeof value !== "undefined") {
+        results.push({ uid, value })
+      }
+    }
+    return await Promise.resolve(results)
+  }
+
+
+  const upsertEntries: UpsertEntries = async (uidsAndValues: UidsAndValues): Promise<void> => {
+    for (const { uid, value } of uidsAndValues) {
+      entryTable[hexEncode(uid)] = value
+    }
+    return await Promise.resolve()
+  }
+
+  const upsertChains: UpsertChains = async (uidsAndValues: UidsAndValues): Promise<void> => {
+    for (const { uid, value } of uidsAndValues) {
+      chainTable[hexEncode(uid)] = value
+    }
+    return await Promise.resolve()
+  }
+
+
+  await upsert(
+    [entryLocation, entryKeyword],
+    searchKey,
+    updateKey,
+    label,
+    fetchEntries,
+    upsertEntries,
+    upsertChains
+  )
+
+  const results = await search(
+    new Set(["ROBERT"]),
+    searchKey,
+    label,
+    100,
+    fetchEntries,
+    fetchChains
+  )
+  expect(results.length).toEqual(1)
+
+  const results2 = await search(
+    new Set(["BOB"]),
+    searchKey,
+    label,
+    100,
+    fetchEntries,
+    fetchChains
+  )
+  expect(results2.length).toEqual(1)
 })
 
 test("in memory", async () => {
@@ -93,7 +166,6 @@ test("in memory", async () => {
 })
 
 test("SQLite", async () => {
-  const sqlite3 = require("sqlite3").verbose()
   const db = new sqlite3.Database(":memory:")
   await new Promise((resolve) => {
     db.run(
@@ -119,7 +191,7 @@ test("SQLite", async () => {
           .join(",")})`,
         uids,
         (err: any, rows: UidsAndValues) => {
-          if (err) reject(err)
+          if (err !== null && typeof err !== "undefined") reject(err)
           resolve(rows)
         }
       )
@@ -135,7 +207,7 @@ test("SQLite", async () => {
           `INSERT OR REPLACE INTO ${table} (uid, value) VALUES(?, ?)`,
           [uid, value],
           (err: any) => {
-            if (err) reject(err)
+            if (err !== null && typeof err !== "undefined") reject(err)
             resolve(null)
           }
         )
@@ -206,15 +278,16 @@ test("Redis", async () => {
   }
 })
 
+// eslint-disable-next-line jsdoc/require-jsdoc
 async function run(
   fetchEntries: FetchEntries,
   fetchChains: FetchChains,
   upsertEntries: UpsertEntries,
   upsertChains: UpsertChains
 ): Promise<void> {
-  const searchKey = new FindexKey(Uint8Array.from(Array(32).keys()))
-  const updateKey = new FindexKey(Uint8Array.from(Array(32).keys()))
-  const label = new Label(Uint8Array.from([1, 2, 3]))
+  const searchKey = new FindexKey(randomBytes(32))
+  const updateKey = new FindexKey(randomBytes(32))
+  const label = new Label(randomBytes(10))
 
   {
     const newIndexedEntries: IndexedEntry[] = []